third_party/WebKit/LayoutTests/http/tests/htmlimports/import-preload-expected.txt

CONSOLE MESSAGE: line 9: PASS: resources/preload.html is preloaded.
CONSOLE WARNING: Styling master document from stylesheets defined in HTML Imports is deprecated, and is planned to be removed in M67, around May 2018. Please refer to https://goo.gl/EGXzpw for possible migration paths.
